fix bug with seq recursion and variant.
the bug: on req parser, the context will be created for each variant after each variant, so we can't just reuse values in context – variant of context's class will be grown as a combination of each variant with itself.

we can crop context in recursion parser and use context as it was before a variant parser.

also clang17 bug: can't recursively call function sometimes – fixed by adding a returning type.

there is a bug with it: crop_ctx method works only with last tag, so we can't use in a deep recursion.