use 1 as value for crop_ctx tag
it was a poitner to ctx for use after crop, but the type was repeating
twice on each crop tag.

seems the clang cannot handle so big types. gcc works normaly.

also reading that type is hard :)

but the problem is still there: the frame tags contains this and creates
bug types