add seq_enable_recursion parser
most of sequence won't use the recursion, so only few seq parsers need to add recursion support in context. the user know wich parser will be used with recursion and can mark it. we also can calcualte it inside (check if recursion parser calls the current seq parser and replace the number with other), but it will increase build time, but the decreasing the build time was a main reason to remove extra info from the context. also with it the user can calculate number more easily (only one or two parser will use it, for example).